t210b01: ers: add "nvidia,active-panel" property
As per the redesign of the driver, active display panel on the given
platform will be selected by property "nvidia,active-panel". Driver
first searches the active panel based on display board-id and if
it doesn't find one, it will look for this property for the
active panel on the given platform.

TDS-1544

Change-Id: I443fc85d4151a8c7aea4a6e6223e22fa351692fb
Signed-off-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-on: http://git-master/r/1328214
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Shu Zhong <shuz@nvidia.com>
GVS: Gerrit_Virtual_Submit