DT: t210: ers: Use pinctrl framework for initialising IO pad voltages
PMC driver register the pinctrl framework to provide the
initialisation of IO pad voltages during init.

Move to use the pinctrl framework to initialise the IO pad
voltages and get rid of the legacy way of initialisation
via pad control driver which is going to be deprecated.

Change-Id: I582dec15b443c304468c839cb3986dc19e4850b0
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1466748
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Venkat Reddy Talla <vreddytalla@nvidia.com>