t210b01: ers: delete nvidia,wake-gpio property in slt sku
Issue: PEX_WAKE line is configured as wake irq in suspend.
It is connected to gpio expander. During suspend it is causing
sc7 failure because it is driven by gpio expander.

Fix: PEX_WAKE line is not connected to PCIe slots, so delete
nvidia,wake-gpio property in slt sku.

bug 1950244

Change-Id: I09be4173f4f5f3c9c20a0ebb4a4c091a4086236f
Signed-off-by: Manikanta <mmaddireddy@nvidia.com>
Reviewed-on: https://git-master/r/1513956
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>